Implement event-based instance replacement
Context:

This change attempts to replace on-demand instances with spot as soon
as they are launched (triggered by the instance entering the pending
state), in order to avoid runnin expensive on-demand capacity in parallel
with the newly launched spot instances. This should cover new groups out
of the box, as well as scaling out or replacement
of terminated spot instances.

It has a few nice side effects concerning cost, since on-demand instances
are likely to run for just a few seconds, as well as potentially tackling
some known issues with ElasticBeanstalk(#343, #344) and deletion of
CloudFormation stacks(#332).

It may also trick the ASG into running the start lifecycle hooks(#296),
and will immediately replace any scaled out instances, partially
addressing #284.

The current/legacy replacement logic will still kick in when enabling
AutoSpotting for the first time on an existing group, as well as when
spot instances fail to be launched so the group must failover to
on-demand capacity in order to avoid downtime.

Current status
- changed regional CloudFormation stack to intercept EC2 instances
as soon as entering the 'pending state'
- implement handling code for this type of events
- implemented initial instance replacement logic leveraging as much as
possible from the existing logic
- atempt populating all the required data structures needed for the
replacement logic
- just got it to compile, the new code wasn't executed yet against real
infrastructure and is likely to crash
- no test coverage for the new code yet

ToDo:
- execute against real infrastructure
- iron out any crashes
- write automated tests for the new code so that coverage doesn't drop
- port regional CloudFormation stack changes to the main template so
it can also run from a StackSet